TTAR-60 Refactoring - Rename Item controller
