+++ src/main/java/com/ttarum/item/controller/ItemController.java
@@ -1,29 +1,23 @@
 package com.ttarum.item.controller;
 
-import com.ttarum.item.domain.Item;
 import com.ttarum.item.dto.response.ItemDetailResponse;
-import com.ttarum.item.service.ItemService;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import static com.ttarum.item.controller.utils.ItemConverter.*;
+@Tag(name = "item", description = "제품")
+public interface ItemController {
 
-@Slf4j
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/items")
-public class ItemController implements IItemController {
-
-    private final ItemService itemService;
-
-    @Override
+    @Operation(summary = "특정 제품의 상세목록")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "성공"),
+            @ApiResponse(responseCode = "400", description = "존재하지 않는 제품")
+    })
+    @Parameter(name = "id", description = "제품의 아이디", example = "1", required = true)
     @GetMapping
-    public ResponseEntity<ItemDetailResponse> getDetail(final Long id) {
-        Item item = itemService.getItem(id);
-        return ResponseEntity.ok(convertToDetailResponse(item));
-    }
+    ResponseEntity<ItemDetailResponse> getDetail(final Long id);
 }
